First attempt at keeping separate setpoints.
Keeping separate heating and cooling setpoints as one would more traditionally expect with a thermostat and then doing magic to give "auto" behavior the treatment that would be expected.

A new attribute "overshootDegrees" is used to figure out how far past the setpoint the temperature is allowed to drift before the unit is set to idle. This can be important in mini-split systems where if units are not idled it can prevent other units in the system from operating the heating/cooling direction they need to.